[NFC] Consolidate JsonProtocolSpecs
